+++ src/main/java/com/ss/carreservation/service/ReservationService.java
@@ -3,6 +3,7 @@ package com.ss.carreservation.service;
 import com.ss.carreservation.dto.ReservationDTO;
 import com.ss.carreservation.entity.Reservation;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -15,7 +16,7 @@ public interface ReservationService {
 
     ReservationDTO getReservationByCarId(Long carId);
 
-    double calculatePrice(Long carId, LocalDateTime start, LocalDateTime end);
+    BigDecimal calculatePrice(Long carId, LocalDateTime start, LocalDateTime end);
 
     ReservationDTO reserveCar(Reservation dto);
 }
